OCR/backend/requirements.txt
Django==5.1.1
djangorestframework==3.16.1
django-cors-headers==4.9.0
djangorestframework-simplejwt==5.5.1
psycopg2-binary==2.9.11
openai==2.16.0
requests==2.32.5
openpyxl==3.1.5
gunicorn==24.1.1
reportlab==4.4.9
python-docx==1.2.0